read description
This is a new dev branch, branch off of this after doing your restore from the troubleshoots. As you work on things and finish them, merge back into dev-branch and ensure that everything is still up and running.

We ought to make front facing webservers, someone to work on the app, and someone to work on the backend. We work together and everything should go smoothly as long as everyone contributes.